Error handling is applied at the backend

+++ controllers/db.js
@@ -44,13 +44,27 @@ export const insertUser = function(req,res) {
 		//"id":req.body.id
 	}
 	let id=uuidv4();
+	if (values["age"]<0 || values["age"]>150){
+		console.log("Age can't be more than 150 and less than 0");
+		res.send("cannot be inserted")
+	}
+	else if (values["gender"]!='m'&&values["gender"]!='M'&&values["gender"]!='f'&&values["gender"]!='F') {
+		console.log("Define gender correctly");
+		res.send("cannot be inserted")
+	}
+	else if(values["name"].length<0||values["name"].length>20) {
+		console.log("Length of the name is exceeded");
+		res.send("Can't be inserted because length of the entered name is exceeded");
+	}
+	else{
 	let sql="insert into users(name,age,gender,id) values ('"+ values["name"] +"'," + values["age"]+ ",'"+ values["gender"]+  "','"+ id+"')";
 	console.log(sql)
 	let query=connection.query(sql,(err, result) => {
-		if (err) throw err;
+		if(err) throw err;
 		console.log("no.of rows inserted into the table : "+result.affectedRows);
 		res.send("no.of rows inserted into the table : "+result.affectedRows)
 	});
+	}
 }
 
 export const getSpeuser = function(req,res) {
@@ -58,7 +72,7 @@ export const getSpeuser = function(req,res) {
 
 	let sql="select name, age, gender from users where id= "+ mysql.escape(id);
 	console.log(sql)
-	let query=connection.query(sql,(err, result) => {
+	let query=connection.query(sql,(err, result,fields) => {
 		if (err) throw err;
 		console.log(result);
 		res.send(result)
@@ -67,13 +81,40 @@ export const getSpeuser = function(req,res) {
 
 export const deleteuser = function(req,res) {
 	const { id }=req.params;
-	let sql="delete from users where id= " +mysql.escape(id);
-	console.log(sql)
-	let query=connection.query(sql,(err, result) => {
-		if (err) throw err;
-		console.log("no.of records deleted: "+result.affectedRows);
-		res.send("no.of records deleted: "+result.affectedRows);
+	let data;
+	function info(callback) {
+		let d_id="select name from users where id= "+mysql.escape(id);
+		let query=connection.query(d_id,(err,result,fields)=> {
+			console.log("resultgh is:"+JSON.stringify(result));
+			if(err) throw err;
+			else {
+				console.log(result.length);
+				console.log("result is:"+JSON.stringify(result));
+				return callback(result);
+				console.log(result.length);
+			}
+		})
+	}
+	info(function(result){
+		//data=result;
+		console.log("data is:"+result.length);
+		// console.log(d_id);
+		if(result.length == 0) {
+			console.log("cannot delete data");
+			res.send("enter valid id");
+		}
+		else {
+			let sql="delete from users where id= " +mysql.escape(id);
+			console.log(sql);
+			let query=connection.query(sql,(err, result) => {
+				if (err) throw err;
+				console.log("no.of records deleted: "+result.affectedRows);
+				res.send("no.of records deleted: "+result.affectedRows);
+			});
+		}
 	});
+	console.log("Testing this...")
+	
 }
 
 export const updateuser = function(req,res) {
@@ -83,7 +124,19 @@ export const updateuser = function(req,res) {
 		"age":req.body.age,
 		"gender":req.body.gender
 	}
-	//const { name, age, gender} = req.body;
+	if (values["age"]<0 || values["age"]>150){
+		console.log("Age can't be more than 150 and less than 0");
+		res.send("cannot be inserted")
+	}
+	else if (values["gender"]!='m'&&values["gender"]!='M'&&values["gender"]!='f'&&values["gender"]!='F') {
+		console.log("Define gender correctly");
+		res.send("cannot be inserted")
+	}
+	else if(values["name"].length<0||values["name"].length>20) {
+		console.log("Length of the name is exceeded");
+		res.send("Can't be inserted because length of the entered name is exceeded");
+	}
+	else {
 	let sql="UPDATE users SET name = '"+ values["name"] +"', age=" + values["age"] + ", gender='"+values["gender"]+"' WHERE id = '"+id+"'";
 	//var sql = "UPDATE users set ? WHERE id = ?";
 	console.log(sql);
@@ -96,6 +149,7 @@ export const updateuser = function(req,res) {
 		console.log("no.of records updated "+result.affectedRows);
 		res.send("no.of records updated "+result.affectedRows);
 	});
+	}
 }
